[CACHE] Reorganize pool implementations: shorter class names (strip CacheItemPool)
