+++ tests/TestSuite/Pimcore/Cache/Core/PdoMysqlCoreHandlerTest.php
@@ -2,7 +2,7 @@
 
 namespace TestSuite\Pimcore\Cache\Core;
 
-use Pimcore\Cache\Pool\PdoMysqlCacheItemPool;
+use Pimcore\Cache\Pool\PdoMysql;
 use TestSuite\Pimcore\Cache\Factory;
 
 class PdoMysqlCoreHandlerTest extends AbstractCoreHandlerTest
@@ -10,7 +10,7 @@ class PdoMysqlCoreHandlerTest extends AbstractCoreHandlerTest
     /**
      * Initializes item pool
      *
-     * @return PdoMysqlCacheItemPool
+     * @return PdoMysql
      */
     protected function createCachePool()
     {
